fix: new row supabase
