+++ package.json
@@ -74,6 +74,7 @@
     "@supabase/supabase-js": "2.33.1",
     "@types/imapflow": "1.0.18",
     "@types/js-yaml": "4.0.9",
+    "@types/pg-format": "1.0.5",
     "@types/showdown": "2.0.6",
     "airtable": "0.11.6",
     "ajv": "8.12.0",
@@ -168,8 +169,9 @@
     "openai": "4.47.1",
     "papaparse": "5.4.1",
     "pdf-parse": "1.1.1",
-    "pdf-text-reader": "5.1.0",
+    "pdf-text-reader": "4.1.0",
     "pg": "8.11.3",
+    "pg-format": "1.0.4",
     "pickleparser": "0.1.0",
     "pino-loki": "2.1.3",
     "posthog-js": "1.83.0",
