+++ packages/pieces/community/pdf/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "dependencies": {
     "@activepieces/pieces-framework": "*",
-    "pdf-text-reader": "5.1.0",
+    "pdf-text-reader": "4.1.0",
     "tslib": "2.6.2"
   },
   "main": "./src/index.js",
