+++ packages/pieces/community/postgres/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@activepieces/piece-postgres",
-  "version": "0.1.5"
+  "version": "0.1.6"
 }
